+crossfade v1.5.0 (#219)
* +play_source cmd: play via mixer

* rename: `TrackEndLoader` -> `RadioInvoker`

depreciate `Players::play` in preference for `players::into_input`

* wip

* wip

* wip

* wip

* break apart the mega file

* rename crates/mockingbird/src/auth.rs crates/mockingbird/src/usersettoken.rs

* fix subgroup check

* wip: cleanup

* revert commit to origin/next-release

* +crossplay: fix audio playing

* fix no play bug

* add debugging info

* trim `use`

* add crossfade command group

* bugging out v1.5.0

* nerd rapper

* fix silly bug

---------
