+++ crates/coggiebot/src/controllers/basic.rs
@@ -39,6 +39,13 @@ async fn rev_cmd(ctx: &Context, msg: &Message) -> CommandResult {
     Ok(())
 }
 
+// #[command]
+// async fn dmsg(ctx: &Context, msg: &Message, args: Args) -> CommandResult {
+//     let x = args.single::<i32>();
+//     msg.channel.delete_messages(&ctx.http, x).await?;
+//     Ok(())
+// }
+
 #[command]
 async fn contribute(ctx: &Context, msg: &Message) -> CommandResult {
     msg.channel_id
