+++ crates/mockingbird/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mockingbird"
-version = "0.0.6"
+version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
@@ -11,6 +11,7 @@ serenity = { version = "0.11", default-features=false, features = ["standard_fra
 tracing = { version = "0.1"}
 tokio = {version = "1.0", default-features=false, features = ["time", "rt"]}
 rand = { version = "0.8" }
+parking_lot = "*"
 
 ####
 serde = { version = "1.0", optional=true }
@@ -18,14 +19,18 @@ serde_json = { version = "1.0", optional=true }
 reqwest = { version = "0.11", optional = true, features = ["cookies"]}
 chrono = {version = "^0.4.26", optional = true }
 cutils = { path = "../cutils", features = ["tokio"], optional=true }
+metrics = { version = "0.22", optional=true }
+
 
 [features]
 default = []
 controller = []
 debug = []
 radio = []
+crossfade = []
+metrics = ["dep:metrics"]
 
-check = ["dep:chrono", "dep:reqwest", "dep:serde", "dep:serde_json"]
+check = ["dep:chrono", "dep:reqwest", "dep:serde", "dep:serde_json", "metrics"]
 ytdl = ["songbird/yt-dlp"]
 deemix = ["dep:serde", "dep:serde_json", "cutils"]
 http-get = ["dep:reqwest"]
