+++ crates/mockingbird/src/lib.rs
@@ -5,6 +5,13 @@ pub mod player;
 #[cfg(feature = "deemix")]
 pub mod deemix;
 
+pub mod events;
+pub mod models;
+pub mod controller;
+pub mod radio;
+pub mod compat;
+pub mod usersettoken;
+pub mod crossfade;
 
 // #[cfg(feature = "http-get")]
 // pub mod httpget;
@@ -20,11 +27,11 @@ pub async fn init(mut cfg: ClientBuilder) -> ClientBuilder {
     tracing::info!("Mockingbird initializing...");
     use songbird::SerenityInit;
 
-
     #[cfg(feature = "controller")]
     {
         use std::collections::HashMap;
-        cfg = cfg.type_map_insert::<player::LazyQueueKey>(HashMap::new());
+        cfg = cfg.type_map_insert::<models::LazyQueueKey>(HashMap::new());
+
     }
 
     cfg.register_songbird()
