+++ flake.nix
@@ -37,6 +37,7 @@
               mockingbird-set-arl-cmd
               mockingbird-http
               mockingbird-radio
+              mockingbird-crossfade
           ]);
 
           coggiebot-stable = cogpkgs.mkCoggiebot {
@@ -88,6 +89,7 @@
           };
 
           packages.cache-target = coggiebot-stable;
+          packages.arange = pkgs.python3.withPackages(ps: with ps; [ numpy ]);
         })) packages;
 
       nixosModules.coggiebot = {pkgs, lib, config, ...}:
